+++ src/assemblyinfo/assemblyinfo.shared.fs
@@ -1,9 +1,9 @@
 ﻿namespace System
 open System.Reflection
 
-[<assembly: AssemblyVersionAttribute("0.0.31")>]
-[<assembly: AssemblyFileVersionAttribute("0.0.31")>]
+[<assembly: AssemblyVersionAttribute("0.0.32")>]
+[<assembly: AssemblyFileVersionAttribute("0.0.32")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.0.31"
+    let [<Literal>] Version = "0.0.32"